grant_code/backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
numpy
packaging==25.0
pandas==2.3.1
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
xlsxwriter==3.2.5
